+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * fix double call of async prompt with async interpreter [#474](https://github.com/jcubic/jquery.terminal/issues/474)
 * fix async + RPC + exec + no system.describe [#475](https://github.com/jcubic/jquery.terminal/issues/475)
 * fix describe === false (exception in js and typscript definition)
+* don't pause terminal when user code return result of calling terminal::read()
 
 ## 2.2.0
 ### Features